Update Deploying an Enterprise Web.md
LAB: Deploying an Enterprise Web
An enterprise intends to deploy its website on HUAWEI CLOUD, and they have the 
following requirements:
1. Database nodes and service nodes are deployed on separate ECSs.
2.ECSs are added or removed as incoming traffic changes over time.
Prerequisites: Log in to HUAWEI CLOUD.Go to the [Lab Desktop] and open Google 
Chrome browser to access the HUAWEI CLOUD login page. Select IAM User Login. In the 
login dialog box, enter the assigned HUAWEI CLOUD lab account and password to log in to 
HUAWEI CLOUD, as shown in the following figure.